refactor(doctor): count severities once for text and JSON output (AI-94)
Text and JSON summaries duplicated issue counting. One shared pass prevents their totals from diverging.
