+++ internal/config/config_test.go
@@ -372,6 +372,44 @@ func TestGetStaleThreshold(t *testing.T) {
 	})
 }
 
+func TestParseDuration(t *testing.T) {
+	tests := []struct {
+		name     string
+		input    string
+		expected time.Duration
+		wantErr  bool
+	}{
+		{"30 days", "30d", 30 * 24 * time.Hour, false},
+		{"2 weeks", "2w", 14 * 24 * time.Hour, false},
+		{"6 months", "6m", 180 * 24 * time.Hour, false},
+		{"1 day", "1d", 24 * time.Hour, false},
+		{"1 week", "1w", 7 * 24 * time.Hour, false},
+		{"1 month", "1m", 30 * 24 * time.Hour, false},
+		{"uppercase D", "30D", 30 * 24 * time.Hour, false},
+		{"uppercase W", "2W", 14 * 24 * time.Hour, false},
+		{"uppercase M", "6M", 180 * 24 * time.Hour, false},
+		{"empty string", "", 0, true},
+		{"invalid format", "abc", 0, true},
+		{"missing number", "d", 0, true},
+		{"invalid unit", "30x", 0, true},
+		{"negative number", "-5d", 0, true},
+		{"zero", "0d", 0, true},
+	}
+
+	for _, tt := range tests {
+		t.Run(tt.name, func(t *testing.T) {
+			got, err := ParseDuration(tt.input)
+			if (err != nil) != tt.wantErr {
+				t.Errorf("ParseDuration(%q) error = %v, wantErr %v", tt.input, err, tt.wantErr)
+				return
+			}
+			if got != tt.expected {
+				t.Errorf("ParseDuration(%q) = %v, want %v", tt.input, got, tt.expected)
+			}
+		})
+	}
+}
+
 func TestGetAutoLockPatterns(t *testing.T) {
 	t.Run("returns defaults when Global is empty", func(t *testing.T) {
 		resetGlobal()
